feat: move Python scripts to src/dotfiles/ package structure
Restructured the codebase to follow standard Python package conventions:
- Moved init.py, swman.py, pkgstatus.py, and logging_config.py to src/dotfiles/
- Updated imports to use relative imports within the package
- Created src/dotfiles/__init__.py with package metadata
- Updated pyproject.toml to point to the new package location
- Maintained all entry points: dotfiles-init, dotfiles-swman, dotfiles-pkgstatus

This change improves code organization and follows Python packaging best practices
while keeping pyproject.toml in the root directory as requested.

All tools tested and working correctly with make test-compile.
