+++ pyproject.toml
@@ -14,15 +14,13 @@ dependencies = [
 ]
 
 [project.scripts]
-dotfiles-init = "init:main"
-dotfiles-swman = "swman:main"
-dotfiles-pkgstatus = "pkgstatus:main"
+dotfiles-init = "dotfiles.init:main"
+dotfiles-swman = "dotfiles.swman:main"
+dotfiles-pkgstatus = "dotfiles.pkgstatus:main"
+
+[tool.hatch.build]
+directory = "src"
 
 [tool.hatch.build.targets.wheel]
-include = [
-    "init.py",
-    "swman.py",
-    "pkgstatus.py",
-    "logging_config.py"
-]
+packages = ["dotfiles"]
 # Trigger PR update
