+++ src/dotfiles/__init__.py
@@ -1,0 +1,3 @@
+"""Dotfiles package for personal system configuration and management tools."""
+
+__version__ = "0.1.0"
