+++ src/dotfiles/swman.py
@@ -19,7 +19,7 @@ from pathlib import Path
 from typing import Dict, List, Tuple
 
 import click
-from logging_config import setup_logging, bind_context, LoggingHelpers, ConsoleOutput
+from .logging_config import setup_logging, bind_context, LoggingHelpers, ConsoleOutput
 
 
 def run_with_streaming_output(
